Change nPeople setter option to boxSize setter
Changing nPeople through terminal arguments has been deemed fallacious design choice as it contradicts the rest of set objectives. In 'test' scenario, number of people was to be hardcoded to 3; in 'file' scenario, number of people depends on the file containing the data; in 'random' scenario, number of people was to be hardcoded to 50. Instead of this, controlling boxSize attribute seems to be much better choice since it never occurs in the program besides the city constructor.